Category info views moved to package 'info'
