+++ src/main/java/egl/client/controller/task/category/CategoryTheoryTaskController.java
@@ -1,7 +1,7 @@
 package egl.client.controller.task.category;
 
 import egl.client.controller.task.TheoryTaskController;
-import egl.client.controller.topic.TranslationsListView;
+import egl.client.controller.topic.info.TranslationsListView;
 import egl.client.model.topic.category.Category;
 import javafx.fxml.FXML;
 import net.rgielen.fxweaver.core.FxmlView;
